+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning][].
 ### Fixed
 
 - Hanging past state of Keystroke input.
+- Custom switch is now disabled when action is default.
 
 ## [2.2.0-beta2]