+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog][],
 and this project adheres to [Semantic Versioning][].
 
+## 2.2.1
+
+### Fixed
+
+- Keys page performance issue
+
 ## [2.2.0]
 
 ### Added